# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Commerce API Documentation",
        "description": "API endpoints for the E-Commerce Application",
        "version": "1.0.0"
    },
    "paths": {
        "/api/addresses": {
            "get": {
                "tags": [
                    "Address"
                ],
                "summary": "Get all addresses for the authenticated user",
                "operationId": "c79dd3c3c64ec5882c4205cbfe71a5c1",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "application/json"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of user addresses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "addresses": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Address"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorUnauthorized"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/address": {
            "post": {
                "tags": [
                    "Address"
                ],
                "summary": "Create a new address for the authenticated user",
                "operationId": "7df30e173742564cc94d94e7199e7bd2",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "application/json"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Address details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "country",
                                    "city",
                                    "street",
                                    "phone_numbers"
                                ],
                                "properties": {
                                    "country": {
                                        "type": "string",
                                        "example": "Egypt"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Giza"
                                    },
                                    "street": {
                                        "type": "string",
                                        "example": "456 Oak Avenue, Building 10"
                                    },
                                    "phone_numbers": {
                                        "description": "Array of phone number strings",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "example": "+201012345678"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Address created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessCreated"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorUnauthorized"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorValidationAddress"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/address/{address}": {
            "put": {
                "tags": [
                    "Address"
                ],
                "summary": "Update an existing address for the authenticated user",
                "operationId": "896da8648439bb541b9f5be395875249",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "application/json"
                        }
                    },
                    {
                        "name": "address",
                        "in": "path",
                        "description": "ID of the address to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "Address details (all fields are optional)",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "country": {
                                        "type": "string",
                                        "example": "Updated Country",
                                        "nullable": true
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "New Cairo",
                                        "nullable": true
                                    },
                                    "street": {
                                        "type": "string",
                                        "example": "789 Pine Road, Floor 5",
                                        "nullable": true
                                    },
                                    "phone_numbers": {
                                        "description": "Array of phone number strings",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "example": "+201298765432"
                                        },
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Address updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessUpdated"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorUnauthorized"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - You do not have permission to update this address",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You do not have permission to update this address."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorNotFound"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorValidationAddress"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Address"
                ],
                "summary": "Delete an address for the authenticated user",
                "operationId": "4dcf64409b91d398e4b4237a930fd910",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "application/json"
                        }
                    },
                    {
                        "name": "address",
                        "in": "path",
                        "description": "ID of the address to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Address deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessDeleted"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorUnauthorized"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - You do not have permission to delete this address",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You do not have permission to delete this address."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorNotFound"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Address": {
                "title": "Address",
                "description": "Address model representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "street": {
                        "type": "string",
                        "example": "123 Example Street"
                    },
                    "city": {
                        "type": "string",
                        "example": "Cairo"
                    },
                    "country": {
                        "type": "string",
                        "example": "Egypt"
                    },
                    "phone_numbers": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "+201012345678"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-11-17T12:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-11-17T12:00:00.000000Z"
                    }
                },
                "type": "object"
            },
...
            "ErrorValidationAddress": {
                "title": "Address Validation Error",
                "description": "Validation error response for address endpoints (create, update address)",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "properties": {
                            "country": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "The country field is required."
                                }
                            },
                            "city": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "The city field is required."
                                }
                            },
                            "street": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "The street must not exceed 255 characters."
                                }
                            },
                            "phone_numbers": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "The phone numbers field must be an array with at least 1 item."
                                }
                            },
                            "phone_numbers.*": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "Each phone number must not exceed 13 characters."
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "ErrorUnauthorized": {
                "title": "Unauthorized",
                "description": "Standard response for 401 Unauthorized",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
...
            "ErrorNotFound": {
                "title": "Not Found",
                "description": "Standard response for 404 Not Found",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "No query results for model [App\\\\Models\\\\ModelName] ID."
                    }
                },
                "type": "object"
            },
...
            "SuccessCreated": {
                "title": "Success Created",
                "description": "Standard success response for 201 Created",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Resource created successfully."
                    },
                    "data": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "SuccessUpdated": {
                "title": "Success Updated",
                "description": "Standard success response for 200 OK (Update)",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Resource updated successfully."
                    },
                    "data": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "SuccessDeleted": {
                "title": "Success Deleted",
                "description": "Standard success response for 200 OK (Delete)",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Resource deleted successfully."
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Sanctum token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}